Refactor assignment process
Tasks support assignee and worker changes how other interact with them.

Tasks, crew, and player all need to account for how that works.

If task has a worker, that worker can be plucked to be a follower. Having a worker means it has an assignee (same entity)

If a task has an assignee, cannot be assigned another assignee, so will reject any attempt